Update dependencies
This updates everything to be compatible with canjs 4/5.
