--- index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>can-connect-ndjson - A Plugin for DoneJS</title>
-    <style>
-      body {
-        width: 800px;
-        margin: 0 auto;
-        text-align: center;
-      }
-      h1 {
-        font-weight: 100;
-        font-size: 60px;
-        color: #393E40;
-      }
-    </style>
-  </head>
-  <body>
-    <h1>Created with the</h1>
-    <a href="https://donejs.com/plugin.html">
-      <img src="https://donejs.com/static/img/donejs-logo-black.svg" alt="DoneJS logo" style="max-width:100%;">
-    </a>
-    <a href="https://www.bitovi.com/blog/create-reusable-components-for-your-donejs-applications">plugin generator</a>
-  </body>
-</html>
