+++ package.json
@@ -28,40 +28,25 @@
     "ci": "npm run build && npm run test && node ./test/test-sauce-labs.js",
     "detect-cycle": "detect-cyclic-packages --ignore done-serve"
   },
-  "main": "can-connect-ndjson",
+  "main": "can-connect-ndjson.js",
   "keywords": [],
-  "steal": {
-    "main": "can-connect-ndjson",
-    "configDependencies": [
-      "live-reload"
-    ],
-    "npmIgnore": [
-      "testee",
-      "generator-donejs",
-      "donejs-cli",
-      "steal-tools"
-    ]
-  },
   "dependencies": {
-    "can-connect": "^1.3.6",
-    "can-define": "^1.0.16",
+    "can-connect": "^3.0.2",
+    "can-define": "^2.5.7",
     "can-namespace": "^1.0.0",
-    "can-ndjson-stream": "^0.1.0",
-    "can-reflect": "^1.2.2",
-    "can-util": "^3.3.0",
-    "mocha": "^3.4.2",
-    "natives": "^1.1.0",
-    "request": "^2.81.0"
+    "can-ndjson-stream": "^1.0.0",
+    "can-reflect": "^1.17.4"
   },
   "devDependencies": {
-    "can-stache": "^3.0.19",
+    "can-stache": "^4.10.5",
     "detect-cyclic-packages": "^1.1.0",
     "http-server": "^0.10.0",
+    "mocha": "^3.4.2",
     "jshint": "^2.9.5",
     "saucelabs": "^1.4.0",
-    "steal": "^1.0.5",
+    "steal": "^2.0.5",
     "steal-qunit": "^1.0.0",
-    "steal-tools": "^1.0.1",
+    "steal-tools": "^2.0.1",
     "test-saucelabs": "0.0.1",
     "testee": "^0.3.2"
   },
