+++ test/test.html
@@ -1,4 +1,5 @@
 <!DOCTYPE html>
 <title>can-connect-ndjson</title>
-<script src="../node_modules/steal/steal.js" main="test/can-connect-ndjson-test"></script>
+<script src="../node_modules/steal/steal.js"
+	main="~/test/can-connect-ndjson-test"></script>
 <div id="qunit-fixture"></div>
